#26 [mod] UI 수정으로 인한 파일 삭제 및 위치 변경
