+++ client/app/src/main/java/com/kaer/menuw/presentation/home/refrigerator/RefrigeratorFragment.kt
@@ -1,4 +1,4 @@
-package com.kaer.menuw.presentation.refrigerator
+package com.kaer.menuw.presentation.home.refrigerator
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,10 +6,10 @@ import android.view.View
 import androidx.fragment.app.activityViewModels
 import com.kaer.menuw.R
 import com.kaer.menuw.databinding.FragmentRefrigeratorBinding
-import com.kaer.menuw.presentation.refrigerator.add.AddIngredientBottomSheet
-import com.kaer.menuw.presentation.refrigerator.add.AddIngredientViewModel
-import com.kaer.menuw.presentation.refrigerator.add.SharedPreferenceManager
-import com.kaer.menuw.presentation.refrigerator.recommend.IngredientRecommendMenuActivity
+import com.kaer.menuw.presentation.home.refrigerator.add.AddIngredientBottomSheet
+import com.kaer.menuw.presentation.home.refrigerator.add.AddIngredientViewModel
+import com.kaer.menuw.presentation.home.refrigerator.add.SharedPreferenceManager
+import com.kaer.menuw.presentation.home.refrigerator.recommend.IngredientRecommendMenuActivity
 import com.kaer.menuw.util.base.BaseDialog
 import com.kaer.menuw.util.base.BaseFragment
