+++ client/app/src/main/java/com/kaer/menuw/presentation/home/refrigerator/add/IngredientListAdapter.kt
@@ -1,35 +1,40 @@
-package com.kaer.menuw.presentation.refrigerator
+package com.kaer.menuw.presentation.home.refrigerator.add
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.kaer.menuw.databinding.ItemIngredientListAddBinding
+import com.kaer.menuw.databinding.ItemIngredientListBinding
 import com.kaer.menuw.domain.entity.IngredientTotal
 import com.kaer.menuw.util.ItemDiffCallback
 import com.kaer.menuw.util.base.BindingAdapter.setCoilImage
 
-class RefrigeratorAdapter :
-    ListAdapter<IngredientTotal.IngredientItem, RefrigeratorAdapter.RefrigeratorViewHolder>(
+class IngredientListAdapter(selectedArray: ArrayList<IngredientTotal.IngredientItem>) :
+    ListAdapter<IngredientTotal.IngredientItem, IngredientListAdapter.IngredientListViewHolder>(
         ItemDiffCallback<IngredientTotal.IngredientItem>(
-            onContentsTheSame = { old, new -> old == new },
-            onItemsTheSame = { old, new -> old.ingredientId == new.ingredientId }
+            onItemsTheSame = { oldItem, newItem -> oldItem.ingredientId == newItem.ingredientId },
+            onContentsTheSame = { oldItem, newItem -> oldItem == newItem }
         )
     ) {
 
     private var onItemClickListener: ((IngredientTotal.IngredientItem) -> Unit)? = null
-    var editEnabled: MutableLiveData<Boolean> = MutableLiveData(false)
-    var selectedIngredientArray: ArrayList<IngredientTotal.IngredientItem> = ArrayList()
 
-    inner class RefrigeratorViewHolder(
-        val binding: ItemIngredientListAddBinding
+    var selectedIngredientArray = selectedArray
+    private val _addEnabled: MutableLiveData<Boolean> = MutableLiveData(false)
+    val addEnabled: LiveData<Boolean>
+        get() = _addEnabled
+
+    inner class IngredientListViewHolder(
+        val binding: ItemIngredientListBinding
     ) : RecyclerView.ViewHolder(binding.root) {
         fun onBind(data: IngredientTotal.IngredientItem, onClickListener: View.OnClickListener) {
             with(binding) {
                 item = data
                 ivIngredientImg.setCoilImage(data.ingredientImageUrl)
+                root.isActivated = selectedIngredientArray.contains(data)
                 root.setOnClickListener(onClickListener)
             }
         }
@@ -43,11 +48,16 @@ class RefrigeratorAdapter :
         return position
     }
 
+    private fun setAddEnabled() {
+        _addEnabled.value = selectedIngredientArray.isNotEmpty()
+    }
+
     private fun ingredientSelection(
-        binding: ItemIngredientListAddBinding,
+        binding: ItemIngredientListBinding,
         ingredientItem: IngredientTotal.IngredientItem
     ) {
-        val isIngredientSelected: Boolean = selectedIngredientArray.contains(ingredientItem)
+        val isIngredientSelected: Boolean =
+            selectedIngredientArray.contains(ingredientItem)
 
         if (isIngredientSelected) {
             removeIngredientItem(
@@ -64,27 +74,26 @@ class RefrigeratorAdapter :
     private fun removeIngredientItem(
         ingredientArray: ArrayList<IngredientTotal.IngredientItem>,
         selectedId: Int,
-        binding: ItemIngredientListAddBinding
+        binding: ItemIngredientListBinding
     ) {
         ingredientArray.removeAt(selectedId)
         binding.root.isActivated = false
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RefrigeratorViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): IngredientListViewHolder {
         val binding =
-            ItemIngredientListAddBinding.inflate(LayoutInflater.from(parent.context), parent, false)
-        return RefrigeratorViewHolder(binding)
+            ItemIngredientListBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        return IngredientListViewHolder(binding)
     }
 
-    override fun onBindViewHolder(holder: RefrigeratorViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: IngredientListViewHolder, position: Int) {
         holder.apply {
             onBind(
                 currentList[position],
                 View.OnClickListener {
-                    if (editEnabled.value == true) {
-                        ingredientSelection(binding, currentList[position])
-                        onItemClickListener?.let { it(currentList[position]) }
-                    }
+                    ingredientSelection(binding, currentList[position])
+                    onItemClickListener?.let { it(currentList[position]) }
+                    setAddEnabled()
                 }
             )
         }
