+++ client/app/src/main/java/com/kaer/menuw/presentation/home/refrigerator/recommend/RecommendMenuAdapter.kt
@@ -1,4 +1,4 @@
-package com.kaer.menuw.presentation.refrigerator.recommend
+package com.kaer.menuw.presentation.home.refrigerator.recommend
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
