--- client/app/src/main/java/com/kaer/menuw/presentation/person/PersonFragment.kt
@@ -1,12 +1,0 @@
-package com.kaer.menuw.presentation.person
-
-import com.kaer.menuw.R
-import com.kaer.menuw.databinding.FragmentPersonBinding
-import com.kaer.menuw.util.base.BaseFragment
-
-class PersonFragment: BaseFragment<FragmentPersonBinding>(R.layout.fragment_person){
-
-    override fun onViewCreated(view: android.view.View, savedInstanceState: android.os.Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-}
