+++ client/app/src/main/res/layout/fragment_refrigerator.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="viewModel"
-            type="com.kaer.menuw.presentation.refrigerator.add.AddIngredientViewModel" />
+            type="com.kaer.menuw.presentation.home.refrigerator.add.AddIngredientViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
